+++ tools/sdk-testgen/packages/autorest.gotest/src/generator/scenarioTestGenerator.ts
@@ -12,8 +12,8 @@ import { GoHelper } from '../util/goHelper';
 import { Helper } from '@autorest/testmodeler/dist/src/util/helper';
 import { MockTestDataRender } from './mockTestGenerator';
 import { OavStepType } from '@autorest/testmodeler/dist/src/common/constant';
+import { OutputVariableModelType, StepRestCallModel, TestDefinitionModel, TestScenarioModel } from '@autorest/testmodeler/dist/src/core/model';
 import { Step } from 'oav/dist/lib/apiScenario/apiScenarioTypes';
-import { StepRestCallModel, TestDefinitionModel, TestScenarioModel } from '@autorest/testmodeler/dist/src/core/model';
 
 export class ScenarioTestDataRender extends MockTestDataRender {
     parentVariables: Record<string, string> = {};
@@ -118,9 +118,24 @@ export class ScenarioTestDataRender extends MockTestDataRender {
                 const example = (step as StepRestCallModel).exampleModel as GoExampleModel;
                 // request and response parse
                 this.fillExampleOutput(example);
+
+                // response output variable convert
                 if (step.outputVariables && Object.keys(step.outputVariables).length > 0) {
-                    this.context.importManager.add('github.com/go-openapi/jsonpointer');
                     example.checkResponse = true;
+                    for (const [variableName, variableConfig] of Object.entries((step as StepRestCallModel).outputVariablesModel)) {
+                        let isPtr = false;
+                        for (let i = 0; i < variableConfig.length; i++) {
+                            if (variableConfig[i].type === OutputVariableModelType.object) {
+                                variableConfig[i]['languageName'] = variableConfig[i].languages.go.name;
+                                isPtr = !variableConfig[i].languages.go?.byValue;
+                            } else if (variableConfig[i].type === OutputVariableModelType.index) {
+                                variableConfig[i]['languageName'] = `[${variableConfig[i].index}]`;
+                            } else if (variableConfig[i].type === OutputVariableModelType.key) {
+                                variableConfig[i]['languageName'] = `["${variableConfig[i].key}"]`;
+                            }
+                        }
+                        step.outputVariables[variableName]['isPtr'] = isPtr;
+                    }
                 }
                 break;
             }